move the deadlock detector test
